+++ CHANGELOG.md
@@ -12,6 +12,8 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ### Added
 
+- New option `--include-legal-hold-membership` on command `code42 users list` that includes the legal hold matter name and ID for any user on legal hold.
+
 - New commands:
   - `code42 users deactivate`
   - `code42 users reactivate`